refactor: homeView displays the newest gratitude and newest anticipation
